+++ src/combinators/endWith.test.js
@@ -1,0 +1,55 @@
+import endWith from './endWith'
+import { mockSignal } from '../emitter'
+
+let s
+let valueSpy, errorSpy, completeSpy
+
+describe('endWith', () => {
+  beforeEach(() => {
+    s = mockSignal()
+
+    valueSpy = jest.fn()
+    errorSpy = jest.fn()
+    completeSpy = jest.fn()
+  })
+
+  it('emits the given value after all other values', () => {
+    endWith(3, s).subscribe(valueSpy, errorSpy, completeSpy)
+
+    expect(valueSpy).not.toHaveBeenCalled()
+    s.value(1)
+    expect(valueSpy).toHaveBeenCalledTimes(1)
+    expect(valueSpy).toHaveBeenLastCalledWith(1)
+    s.value(2)
+    expect(valueSpy).toHaveBeenCalledTimes(2)
+    expect(valueSpy).toHaveBeenLastCalledWith(2)
+    s.complete()
+    expect(valueSpy).toHaveBeenCalledTimes(3)
+    expect(valueSpy).toHaveBeenLastCalledWith(3)
+  })
+
+  it('emits an error when the given signal emits an error', () => {
+    endWith(0, s).subscribe(valueSpy, errorSpy, completeSpy)
+
+    expect(errorSpy).not.toHaveBeenCalled()
+    s.error('foo')
+    expect(errorSpy).toHaveBeenCalledTimes(1)
+    expect(errorSpy).toHaveBeenCalledWith('foo')
+  })
+
+  it('completes when the given signal is completed', () => {
+    endWith(0, s).subscribe(valueSpy, errorSpy, completeSpy)
+
+    expect(completeSpy).not.toHaveBeenCalled()
+    s.complete()
+    expect(completeSpy).toHaveBeenCalledTimes(1)
+  })
+
+  it('unmounts the given signal when the returned signal is unsubscribed', () => {
+    const a = endWith(0, s).subscribe()
+
+    expect(s.unmount).not.toHaveBeenCalled()
+    a.unsubscribe()
+    expect(s.unmount).toHaveBeenCalledTimes(1)
+  })
+})